Move these to the client and use a Mock'd provider as we are just testing the contract of the methods based on state.

+++ qpid-jms-client/src/test/java/org/apache/qpid/jms/test/Wait.java
@@ -21,7 +21,7 @@ import java.util.concurrent.TimeUnit;
 public class Wait {
 
     public static final long MAX_WAIT_MILLIS = 30 * 1000;
-    public static final int SLEEP_MILLIS = 1000;
+    public static final long SLEEP_MILLIS = 1000;
 
     public interface Condition {
         boolean isSatisified() throws Exception;
@@ -35,7 +35,7 @@ public class Wait {
         return waitFor(condition, duration, SLEEP_MILLIS);
     }
 
-    public static boolean waitFor(final Condition condition, final long duration, final int sleepMillis) throws Exception {
+    public static boolean waitFor(final Condition condition, final long duration, final long sleepMillis) throws Exception {
 
         final long expiry = System.currentTimeMillis() + duration;
         boolean conditionSatisified = condition.isSatisified();
